Split out test generators into a seperate file
